+++ bili/aether/attacks/injector.py
@@ -28,7 +28,7 @@ Blocking vs. fire-and-forget:
 import datetime
 import logging
 import uuid
-from concurrent.futures import ThreadPoolExecutor
+from concurrent.futures import Future, ThreadPoolExecutor
 from pathlib import Path
 from typing import TYPE_CHECKING, Optional
 
@@ -48,6 +48,7 @@ _STRATEGY_MAP = {
     AttackType.MEMORY_POISONING: "inject_memory_poisoning",
     AttackType.AGENT_IMPERSONATION: "inject_agent_impersonation",
     AttackType.BIAS_INHERITANCE: "inject_bias_inheritance",
+    AttackType.JAILBREAK: "inject_prompt_injection",
 }
 
 
@@ -149,7 +150,7 @@ class AttackInjector:
         )
 
         if not blocking:
-            self._thread_pool.submit(
+            future = self._thread_pool.submit(
                 self._run_attack,
                 attack_id,
                 agent_id,
@@ -159,6 +160,7 @@ class AttackInjector:
                 now,
                 track_propagation,
             )
+            future.add_done_callback(self._log_future_exception)
             LOGGER.info(
                 "AttackInjector: submitted non-blocking attack %s for agent '%s'",
                 attack_id,
@@ -324,9 +326,15 @@ class AttackInjector:
             attack_type=attack_type.value,
         )
 
+    def _log_future_exception(self, future: Future) -> None:
+        """Done-callback: log any exception raised by a background attack thread."""
+        exc = future.exception()
+        if exc:
+            LOGGER.error("AttackInjector: background attack thread failed: %s", exc)
+
     def close(self) -> None:
-        """Shut down the background thread pool."""
-        self._thread_pool.shutdown(wait=False)
+        """Shut down the background thread pool, waiting for pending attacks to finish."""
+        self._thread_pool.shutdown(wait=True)
 
     def __enter__(self) -> "AttackInjector":
         return self
