+++ bili/aether/attacks/models.py
@@ -32,6 +32,7 @@ class AttackType(str, Enum):  # pylint: disable=too-few-public-methods
     MEMORY_POISONING = "memory_poisoning"
     AGENT_IMPERSONATION = "agent_impersonation"
     BIAS_INHERITANCE = "bias_inheritance"
+    JAILBREAK = "jailbreak"
 
 
 class AgentObservation(BaseModel):
@@ -55,7 +56,7 @@ class AgentObservation(BaseModel):
     received_payload: bool
     influenced: bool
     output_excerpt: Optional[str] = None
-    resisted: bool
+    resisted: bool = False
 
     @model_validator(mode="after")
     def _enforce_resisted(self) -> "AgentObservation":
