+++ bili/aether/evaluator/__init__.py
@@ -3,8 +3,18 @@
 Usage::
 
     from bili.aether.evaluator import SemanticEvaluator, VerdictResult
+    from bili.aether.evaluator import JAILBREAK_JUDGE_PROMPT, JAILBREAK_SCORE_DESCRIPTIONS
 """
 
+from bili.aether.evaluator.evaluator_config import (
+    JAILBREAK_JUDGE_PROMPT,
+    JAILBREAK_SCORE_DESCRIPTIONS,
+)
 from bili.aether.evaluator.semantic_evaluator import SemanticEvaluator, VerdictResult
 
-__all__ = ["SemanticEvaluator", "VerdictResult"]
+__all__ = [
+    "SemanticEvaluator",
+    "VerdictResult",
+    "JAILBREAK_JUDGE_PROMPT",
+    "JAILBREAK_SCORE_DESCRIPTIONS",
+]
