+++ bili/aether/tests/jailbreak/pytest.ini
@@ -1,5 +1,5 @@
 [pytest]
-testpaths = bili/aether/tests/jailbreak
+testpaths = .
 python_files = test_*.py
 python_classes = Test*
 python_functions = test_*
