+++ bili/aether/tests/jailbreak/run_jailbreak_suite.py
@@ -43,9 +43,6 @@ Results are written to:
 """
 
 import argparse
-import csv
-import datetime
-import json
 import logging
 import sys
 from pathlib import Path
@@ -72,33 +69,14 @@ _REPO_ROOT = _find_repo_root()
 if str(_REPO_ROOT) not in sys.path:
     sys.path.insert(0, str(_REPO_ROOT))
 
-from bili.aether.attacks.injector import (  # noqa: E402  pylint: disable=wrong-import-position
-    AttackInjector,
-)
 from bili.aether.attacks.models import (  # noqa: E402  pylint: disable=wrong-import-position
     InjectionPhase,
 )
-from bili.aether.config.loader import (  # noqa: E402  pylint: disable=wrong-import-position
-    load_mas_from_yaml,
-)
-from bili.aether.evaluator.evaluator_config import (  # noqa: E402  pylint: disable=wrong-import-position
-    JAILBREAK_JUDGE_PROMPT,
-    JAILBREAK_SCORE_DESCRIPTIONS,
-)
-from bili.aether.runtime.executor import (  # noqa: E402  pylint: disable=wrong-import-position
-    MASExecutor,
-)
-from bili.aether.security.detector import (  # noqa: E402  pylint: disable=wrong-import-position
-    SecurityEventDetector,
-)
-from bili.aether.security.logger import (  # noqa: E402  pylint: disable=wrong-import-position
-    SecurityEventLogger,
-)
 from bili.aether.tests._helpers import (  # noqa: E402  pylint: disable=wrong-import-position
-    config_fingerprint as _config_fingerprint_helper,
+    CONFIG_PATHS,
 )
-from bili.aether.tests.injection.payloads.prompt_injection_payloads import (  # noqa: E402  pylint: disable=wrong-import-position
-    InjectionPayload,
+from bili.aether.tests._suite_runner import (  # noqa: E402  pylint: disable=wrong-import-position
+    run_suite,
 )
 from bili.aether.tests.jailbreak.payloads.jailbreak_payloads import (  # noqa: E402  pylint: disable=wrong-import-position
     JAILBREAK_PAYLOADS,
@@ -106,376 +84,13 @@ from bili.aether.tests.jailbreak.payloads.jailbreak_payloads import (  # noqa: E
 
 LOGGER = logging.getLogger(__name__)
 
-# ---------------------------------------------------------------------------
-# Configuration
-# ---------------------------------------------------------------------------
-
 _JAILBREAK_DIR = Path(__file__).parent
 _RESULTS_DIR = _JAILBREAK_DIR / "results"
-
-_CONFIG_PATHS: list[str] = [
-    "bili/aether/config/examples/simple_chain.yaml",
-    "bili/aether/config/examples/hierarchical_voting.yaml",
-    "bili/aether/config/examples/supervisor_moderation.yaml",
-    "bili/aether/config/examples/consensus_network.yaml",
-    "bili/aether/config/examples/custom_escalation.yaml",
-]
 
 _ALL_PHASES: list[str] = [
     InjectionPhase.PRE_EXECUTION.value,
     InjectionPhase.MID_EXECUTION.value,
 ]
-
-_CSV_COLUMNS: list[str] = [
-    "payload_id",
-    "injection_type",
-    "severity",
-    "stub_mode",
-    "mas_id",
-    "phase",
-    "tier1_pass",
-    "tier2_influenced",
-    "tier2_resistant",
-    "tier3_score",
-    "tier3_confidence",
-    "tier3_reasoning",
-    "attack_suite",
-]
-
-# ---------------------------------------------------------------------------
-# Helpers
-# ---------------------------------------------------------------------------
-
-
-def _config_fingerprint(config, yaml_path: str) -> dict:
-    """Thin wrapper that forwards to the shared helper with the module _REPO_ROOT."""
-    return _config_fingerprint_helper(config, yaml_path, _REPO_ROOT)
-
-
-def _target_agent_id(config) -> str:
-    """Return the injection target agent for a MAS config.
-
-    Uses ``config.entry_point`` when set (e.g. supervisor workflows); falls
-    back to the first agent in the config's agent list.
-
-    Raises:
-        ValueError: If the config has no agents defined.
-    """
-    if not config.agents:
-        raise ValueError(
-            f"MAS config '{config.mas_id}' has no agents defined — "
-            "cannot determine injection target"
-        )
-    if config.entry_point:
-        return config.entry_point
-    return config.agents[0].agent_id
-
-
-def _build_result_dict(
-    ip: InjectionPayload,
-    config,
-    yaml_path: str,
-    phase: str,
-    attack_result,
-    stub_mode: bool,
-    tier3_rows: list | None,
-) -> dict:
-    """Assemble the full result JSON dict for one test case."""
-    # Tier 3: aggregate score across agents (max score = most influenced)
-    tier3_score: str = ""
-    tier3_confidence: str = ""
-    tier3_reasoning: str = ""
-
-    if tier3_rows:
-        # Use the highest-scoring verdict (most concerning finding)
-        best = max(tier3_rows, key=lambda v: v.score)
-        tier3_score = str(best.score)
-        tier3_confidence = best.confidence
-        tier3_reasoning = best.reasoning
-
-    resistant_list = sorted(attack_result.resistant_agents)
-
-    return {
-        "payload_id": ip.payload_id,
-        "injection_type": ip.injection_type,
-        "severity": ip.severity,
-        "mas_id": config.mas_id,
-        "injection_phase": phase,
-        "attack_suite": "jailbreak",
-        "config_fingerprint": _config_fingerprint(config, yaml_path),
-        "execution": {
-            "success": attack_result.success,
-            "duration_ms": (
-                (attack_result.completed_at - attack_result.injected_at).total_seconds()
-                * 1000
-                if attack_result.completed_at
-                else 0.0
-            ),
-            "agent_count": len(attack_result.propagation_path),
-            "message_count": 0,  # not tracked at injection level
-        },
-        "propagation_path": attack_result.propagation_path,
-        "influenced_agents": attack_result.influenced_agents,
-        "resistant_agents": resistant_list,
-        "run_metadata": {
-            "timestamp": datetime.datetime.now(datetime.timezone.utc).isoformat(),
-            "stub_mode": stub_mode,
-            "semantic_tier": "skipped" if stub_mode else "evaluated",
-            "tier3_score": tier3_score,
-            "tier3_confidence": tier3_confidence,
-            "tier3_reasoning": tier3_reasoning,
-        },
-    }
-
-
-def _write_result(result_dict: dict) -> Path:
-    """Write one result JSON to results/{mas_id}/{payload_id}_{phase}.json."""
-    out_dir = _RESULTS_DIR / result_dict["mas_id"]
-    out_dir.mkdir(parents=True, exist_ok=True)
-    filename = f"{result_dict['payload_id']}_{result_dict['injection_phase']}.json"
-    out_path = out_dir / filename
-    out_path.write_text(json.dumps(result_dict, indent=2), encoding="utf-8")
-    return out_path
-
-
-def _write_csv(rows: list[dict]) -> Path:
-    """Write the results matrix CSV to results/jailbreak_results_matrix.csv.
-
-    The CSV is fully RFC 4180-compliant — no comment lines are prepended.
-    Stub mode is encoded as the ``stub_mode`` column value instead.
-    """
-    _RESULTS_DIR.mkdir(parents=True, exist_ok=True)
-    csv_path = _RESULTS_DIR / "jailbreak_results_matrix.csv"
-
-    with csv_path.open("w", newline="", encoding="utf-8") as fh:
-        writer = csv.DictWriter(fh, fieldnames=_CSV_COLUMNS)
-        writer.writeheader()
-        writer.writerows(rows)
-
-    return csv_path
-
-
-def _print_summary(matrix_rows: list[dict]) -> None:
-    """Print a compact summary table to stdout."""
-    total = len(matrix_rows)
-    passed = sum(1 for r in matrix_rows if r["tier1_pass"] is True)
-    influenced_any = sum(1 for r in matrix_rows if r["tier2_influenced"] != "[]")
-
-    print("\n" + "=" * 60)
-    print("Jailbreak Suite Summary")
-    print("=" * 60)
-    print(f"  Total test cases : {total}")
-    print(f"  Tier 1 pass      : {passed}/{total}")
-    print(f"  Tier 2 influenced: {influenced_any} cases had ≥1 influenced agent")
-    print("=" * 60 + "\n")
-
-
-# ---------------------------------------------------------------------------
-# Per-config runner
-# ---------------------------------------------------------------------------
-
-
-def _run_config(  # pylint: disable=too-many-locals,too-many-arguments
-    yaml_path: str,
-    payloads: list[InjectionPayload],
-    phases: list[str],
-    stub_mode: bool,
-    semantic_evaluator,
-    baseline_results_dir: Path | None,
-) -> list[dict]:
-    """Run all payloads × phases for one MAS config.
-
-    Returns a list of CSV matrix row dicts.
-    """
-    full_path = _REPO_ROOT / yaml_path
-    if not full_path.exists():
-        print(f"  Config not found, skipping: {yaml_path}", file=sys.stderr)
-        return []
-
-    config = load_mas_from_yaml(str(full_path))
-    if stub_mode:
-        for agent in config.agents:
-            agent.model_name = None
-
-    mas_id = config.mas_id
-    target_id = _target_agent_id(config)
-
-    # Per-config log paths (shared across all payloads in this config)
-    config_results_dir = _RESULTS_DIR / mas_id
-    config_results_dir.mkdir(parents=True, exist_ok=True)
-    attack_log_path = config_results_dir / "attack_log.ndjson"
-    sec_log_path = config_results_dir / "security_events.ndjson"
-
-    sec_logger = SecurityEventLogger(log_path=sec_log_path)
-    detector = SecurityEventDetector(
-        logger=sec_logger,
-        attack_log_path=attack_log_path,
-    )
-
-    # Only initialise the outer MASExecutor when mid_execution is requested.
-    # For pre_execution-only runs, AttackInjector creates its own fresh
-    # executor internally — this outer instance would never be used.
-    needs_executor = InjectionPhase.MID_EXECUTION.value in phases
-    executor = None
-    if needs_executor:
-        executor = MASExecutor(config)
-        executor.initialize()
-
-    matrix_rows: list[dict] = []
-    print(
-        f"\n[{mas_id}] target_agent={target_id!r}  "
-        f"({len(payloads)} payloads × {len(phases)} phases)"
-    )
-
-    with AttackInjector(
-        config,
-        executor,
-        log_path=attack_log_path,
-        security_detector=detector,
-    ) as injector:
-        for ip in payloads:
-            for phase in phases:
-                label = f"  {ip.payload_id} / {phase}"
-                print(f"{label} ... ", end="", flush=True)
-
-                try:
-                    attack_result = injector.inject_attack(
-                        agent_id=target_id,
-                        attack_type="prompt_injection",
-                        payload=ip.payload,
-                        injection_phase=phase,
-                        blocking=True,
-                        track_propagation=True,
-                    )
-                except Exception as exc:  # pylint: disable=broad-exception-caught
-                    LOGGER.error(
-                        "inject_attack failed for %s / %s / %s: %s",
-                        mas_id,
-                        ip.payload_id,
-                        phase,
-                        exc,
-                    )
-                    print(f"ERROR: {exc}")
-                    matrix_rows.append(
-                        {
-                            "payload_id": ip.payload_id,
-                            "injection_type": ip.injection_type,
-                            "severity": ip.severity,
-                            "stub_mode": stub_mode,
-                            "mas_id": mas_id,
-                            "phase": phase,
-                            "tier1_pass": False,
-                            "tier2_influenced": "[]",
-                            "tier2_resistant": "[]",
-                            "tier3_score": "",
-                            "tier3_confidence": "",
-                            "tier3_reasoning": str(exc),
-                            "attack_suite": "jailbreak",
-                        }
-                    )
-                    continue
-
-                # Tier 3 — semantic evaluation (skipped in stub mode)
-                tier3_rows = None
-                if not stub_mode and semantic_evaluator is not None:
-                    baseline_dict = _load_baseline(baseline_results_dir, mas_id)
-                    if baseline_dict is not None:
-                        try:
-                            tier3_rows = semantic_evaluator.evaluate(
-                                baseline_result=baseline_dict,
-                                attack_result=attack_result,
-                            )
-                        except (
-                            Exception
-                        ) as t3_exc:  # pylint: disable=broad-exception-caught
-                            LOGGER.warning(
-                                "SemanticEvaluator failed for %s/%s: %s",
-                                ip.payload_id,
-                                phase,
-                                t3_exc,
-                            )
-
-                result_dict = _build_result_dict(
-                    ip=ip,
-                    config=config,
-                    yaml_path=yaml_path,
-                    phase=phase,
-                    attack_result=attack_result,
-                    stub_mode=stub_mode,
-                    tier3_rows=tier3_rows,
-                )
-
-                out_path = _write_result(result_dict)
-                status = "ok" if attack_result.success else "FAIL"
-                influenced_count = len(attack_result.influenced_agents)
-                print(f"{status}  (influenced={influenced_count}) → {out_path.name}")
-
-                influenced_json = json.dumps(sorted(attack_result.influenced_agents))
-                resistant_json = json.dumps(sorted(attack_result.resistant_agents))
-
-                matrix_rows.append(
-                    {
-                        "payload_id": ip.payload_id,
-                        "injection_type": ip.injection_type,
-                        "severity": ip.severity,
-                        "stub_mode": stub_mode,
-                        "mas_id": mas_id,
-                        "phase": phase,
-                        "tier1_pass": attack_result.success,
-                        "tier2_influenced": influenced_json,
-                        "tier2_resistant": resistant_json,
-                        "tier3_score": result_dict["run_metadata"]["tier3_score"],
-                        "tier3_confidence": result_dict["run_metadata"][
-                            "tier3_confidence"
-                        ],
-                        "tier3_reasoning": result_dict["run_metadata"][
-                            "tier3_reasoning"
-                        ],
-                        "attack_suite": "jailbreak",
-                    }
-                )
-
-    return matrix_rows
-
-
-def _load_baseline(
-    baseline_results_dir: Path | None,
-    mas_id: str,
-) -> dict | None:
-    """Try to load a baseline result for Tier 3 comparison.
-
-    Baseline results live at ``baseline_results_dir/{mas_id}/*.json``.
-    Jailbreak payloads have no 1:1 mapping to baseline prompts; the first
-    available baseline result for this config (alphabetically) is used as
-    the comparison reference.  If no baseline file exists, returns ``None``
-    — the evaluator will still score using the injected output alone with
-    lower confidence.
-
-    Args:
-        baseline_results_dir: Path to the baseline results root, or ``None``.
-        mas_id:               MAS identifier to look up.
-
-    Returns:
-        Parsed baseline result dict, or ``None`` if unavailable.
-    """
-    if baseline_results_dir is None:
-        return None
-    config_baseline_dir = baseline_results_dir / mas_id
-    if not config_baseline_dir.exists():
-        return None
-    result_files = sorted(config_baseline_dir.glob("*.json"))
-    if not result_files:
-        return None
-    try:
-        return json.loads(result_files[0].read_text(encoding="utf-8"))
-    except (OSError, json.JSONDecodeError) as exc:
-        LOGGER.warning("Could not load baseline for %s: %s", mas_id, exc)
-        return None
-
-
-# ---------------------------------------------------------------------------
-# Main
-# ---------------------------------------------------------------------------
 
 
 def main() -> None:
@@ -497,7 +112,7 @@ def main() -> None:
     parser.add_argument(
         "--configs",
         nargs="+",
-        default=_CONFIG_PATHS,
+        default=CONFIG_PATHS,
         metavar="YAML_PATH",
         help="Override the list of YAML config paths to run.",
     )
@@ -560,6 +175,8 @@ def main() -> None:
     if not args.stub:
         try:
             from bili.aether.evaluator import (  # pylint: disable=import-outside-toplevel
+                JAILBREAK_JUDGE_PROMPT,
+                JAILBREAK_SCORE_DESCRIPTIONS,
                 SemanticEvaluator,
             )
 
@@ -570,28 +187,20 @@ def main() -> None:
         except (ImportError, RuntimeError) as exc:
             LOGGER.warning("Could not initialise SemanticEvaluator: %s", exc)
 
-    # Run all configs
-    all_matrix_rows: list[dict] = []
-    for yaml_path in args.configs:
-        rows = _run_config(
-            yaml_path=yaml_path,
-            payloads=payloads,
-            phases=args.phases,
-            stub_mode=args.stub,
-            semantic_evaluator=semantic_evaluator,
-            baseline_results_dir=baseline_results_dir,
-        )
-        all_matrix_rows.extend(rows)
-
-    # Write CSV
-    if all_matrix_rows:
-        csv_path = _write_csv(all_matrix_rows)
-        _print_summary(all_matrix_rows)
-        print(f"Results matrix written to: {csv_path}")
-
-    total = len(all_matrix_rows)
-    passed = sum(1 for r in all_matrix_rows if r["tier1_pass"] is True)
-    sys.exit(0 if passed == total else 1)
+    run_suite(
+        payloads=payloads,
+        attack_suite="jailbreak",
+        attack_type="jailbreak",
+        csv_filename="jailbreak_results_matrix.csv",
+        suite_name="Jailbreak Suite",
+        results_dir=_RESULTS_DIR,
+        repo_root=_REPO_ROOT,
+        config_paths=args.configs,
+        phases=args.phases,
+        stub=args.stub,
+        semantic_evaluator=semantic_evaluator,
+        baseline_results_dir=baseline_results_dir,
+    )
 
 
 if __name__ == "__main__":
